Setup testing files
